fixed bug with global state
