+++ code/src/Components/ToDoItem.js
@@ -1,25 +1,15 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { useSelector, useDispatch } from 'react-redux'
-import { todoitem } from './Reducers/todoitem';
-import { Checkmark } from './Checkmark'
+import { ItemText } from './ItemText'
 
 export const ToDoItem = () => {
-  const dispatch = useDispatch();
+
   const allItems = useSelector((store) => store.todoitem.items)
-  const [done, setDone] = useState(false)
-
+  
   return (
-
-
     allItems.map(item => {
       return (
-        <section className={`todo-item ${done ? "done" : ""}`}>
-          <div className={`finished-status ${done ? "fill" : ""}`} onClick={() => setDone(!done)}>
-            {done ? < Checkmark /> : ""}
-          </div>
-          <p>{item.name}</p>
-          <button onClick={() => { dispatch(todoitem.actions.removeItem(item)) }}>remove</button>
-        </section>
+        < ItemText item={item}/>
       )
     })
