+++ code/src/index.css
@@ -63,13 +63,14 @@ form button {
 form input {
   border: none;
   border: none;
-  background-color: #eefbff45;
+  background-color: transparent;
   font-size: 1.5em;
   height: 100%;
   width: 90%;
 }
 
 .finished-status {
+  position: relative;
   height: 25px;
   width: 25px;
   margin: 0 5px;
@@ -77,11 +78,22 @@ form input {
   border-radius: 50%;
 }
 
+ul {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.todo-item input {
+  opacity: 0;
+  position: absolute;
+}
+
 .todo-item {
   display: flex;
   align-items: center;
   height: 55px;
-  background-color: #eefbff45;
+  background-color: #add8e645;
   border-bottom: solid 1px grey;
   font-size: 1.5em;
 }
@@ -97,4 +109,5 @@ form input {
 
 .done {
   color: #D0C9D6;
+  background-color: white;
 }
